session_state_management/basic_stateful_session.py
import uuid

from dotenv import load_dotenv
from google.adk.runners import Runner
from google.adk.sessions import InMemorySessionService
from google.genai import types
from question_answer_agent import question_answer_agent
import asyncio

load_dotenv()


# Create a new session service to store state
session_service_stateful = InMemorySessionService()

initial_state = {
    "user_name": "Tony Stark (Iron Man)",
    "user_preferences": """
        Enjoys inventing new technology and building advanced suits of armor.
        Loves shawarma and fine dining.
        Favorite TV show is anything with science or engineering.
        Enjoys witty banter and being the center of attention.
        Has a soft spot for helping others, even if he hides it behind sarcasm.
    """,
}

async def main():
    # Create a NEW session
    APP_NAME = "JARVIS"
    USER_ID = "tonystark"
    SESSION_ID = str(uuid.uuid4())
    stateful_session = await session_service_stateful.create_session(
        app_name=APP_NAME,
        user_id=USER_ID,
        session_id=SESSION_ID,
        state=initial_state,
    )
    print("CREATED NEW SESSION:")
    print(f"\tSession ID: {SESSION_ID}")

    runner = Runner(
        agent=question_answer_agent,
        app_name=APP_NAME,
        session_service=session_service_stateful,
    )

    new_message = types.Content(
        role="user", parts=[types.Part(text="What is Tony stark's favorite genre of TV show ?")]
    )

    for event in runner.run(
        user_id=USER_ID,
        session_id=SESSION_ID,
        new_message=new_message,
    ):
        if event.is_final_response():
            if event.content and event.content.parts:
                print(f"Final Response: {event.content.parts[0].text}")

    print("==== Session Event Exploration ====")
    session = await session_service_stateful.get_session(
        app_name=APP_NAME, user_id=USER_ID, session_id=SESSION_ID
    )

    # Log final Session state
    print("=== Final Session State ===")
    for key, value in session.state.items():
        print(f"{key}: {value}")

if __name__ == "__main__":
    asyncio.run(main())


# cd session_state_management
# python basic_stateful_session.py

